Fix latest issues from the Demo

+++ src/Warehouse/Application/Features/Queries/GetProducts/GetProductsQuery.cs
@@ -4,7 +4,7 @@ using MediatR;
 namespace Application.Features.Queries.GetProducts;
 
 public record GetProductsQuery(
-    double? MinPrice,
-    double? MaxPrice,
+    decimal? MinPrice,
+    decimal? MaxPrice,
     List<string> Highlights,
     string Size) : IRequest<GetProductsDto>;
