+++ src/Warehouse/Domain/Models/ProductsSummary.cs
@@ -2,9 +2,9 @@
 
 public class ProductsSummary
 {
-    public double MinPrice { get; set; }
+    public decimal MinPrice { get; set; }
 
-    public double MaxPrice { get; set; }
+    public decimal MaxPrice { get; set; }
 
     public List<string> Sizes { get; set; }
