+++ src/Warehouse/WebApi/Models/Requests/GetFilteredProductsRequest.cs
@@ -2,9 +2,9 @@
 
 public class GetFilteredProductsRequest
 {
-    public double? MinPrice { get; set; }
+    public decimal? MinPrice { get; set; }
 
-    public double? MaxPrice { get; set; }
+    public decimal? MaxPrice { get; set; }
 
     public List<string> Highlights { get; set; }
