+++ src/Warehouse/WebApi/Models/Responses/ProductsSummaryResponse.cs
@@ -2,9 +2,9 @@
 
 public class ProductsSummaryResponse
 {
-    public double MinPrice { get; set; }
+    public decimal MinPrice { get; set; }
 
-    public double MaxPrice { get; set; }
+    public decimal MaxPrice { get; set; }
 
     public List<string> Sizes { get; set; }
